+++ crates/ruma-server-util/src/lib.rs
@@ -4,3 +4,4 @@
 
 #![warn(missing_docs)]
 pub mod authorization;
+pub mod media;